+++ src/Wrapper/wrapper-string.h
@@ -6,4 +6,4 @@
 
 void wrapper_string_trim_right(TCHAR* chars);
 int wrapper_string_duplicate(TCHAR** result, TCHAR* source, wrapper_error_t** error);
-void wrapper_string_copy(TCHAR* destination, const int destination_max_size, TCHAR* source);
+void wrapper_string_copy(TCHAR* destination, const size_t destination_max_size, TCHAR* source);
